+++ docs-website/src/components/icons/InstallationIcon.jsx
@@ -1,19 +1,11 @@
-import { DarkMode, Gradient, LightMode } from '@/components/Icon'
+import { DarkMode, Gradient, LightMode } from '@/components/Icon';
 
 export function InstallationIcon({ id, color }) {
 	return (
 		<>
 			<defs>
-				<Gradient
-					id={`${id}-gradient`}
-					color={color}
-					gradientTransform="matrix(0 21 -21 0 12 3)"
-				/>
-				<Gradient
-					id={`${id}-gradient-dark`}
-					color={color}
-					gradientTransform="matrix(0 21 -21 0 16 7)"
-				/>
+				<Gradient id={`${id}-gradient`} color={color} gradientTransform="matrix(0 21 -21 0 12 3)" />
+				<Gradient id={`${id}-gradient-dark`} color={color} gradientTransform="matrix(0 21 -21 0 16 7)" />
 			</defs>
 			<LightMode>
 				<circle cx={12} cy={12} r={12} fill={`url(#${id}-gradient)`} />
@@ -37,5 +29,5 @@ export function InstallationIcon({ id, color }) {
 				/>
 			</DarkMode>
 		</>
-	)
+	);
 }