+++ docs-website/src/components/icons/PluginsIcon.jsx
@@ -1,24 +1,12 @@
-import { DarkMode, Gradient, LightMode } from '@/components/Icon'
+import { DarkMode, Gradient, LightMode } from '@/components/Icon';
 
 export function PluginsIcon({ id, color }) {
 	return (
 		<>
 			<defs>
-				<Gradient
-					id={`${id}-gradient`}
-					color={color}
-					gradientTransform="matrix(0 21 -21 0 20 11)"
-				/>
-				<Gradient
-					id={`${id}-gradient-dark-1`}
-					color={color}
-					gradientTransform="matrix(0 22.75 -22.75 0 16 6.25)"
-				/>
-				<Gradient
-					id={`${id}-gradient-dark-2`}
-					color={color}
-					gradientTransform="matrix(0 14 -14 0 16 10)"
-				/>
+				<Gradient id={`${id}-gradient`} color={color} gradientTransform="matrix(0 21 -21 0 20 11)" />
+				<Gradient id={`${id}-gradient-dark-1`} color={color} gradientTransform="matrix(0 22.75 -22.75 0 16 6.25)" />
+				<Gradient id={`${id}-gradient-dark-2`} color={color} gradientTransform="matrix(0 14 -14 0 16 10)" />
 			</defs>
 			<LightMode>
 				<circle cx={20} cy={20} r={12} fill={`url(#${id}-gradient)`} />
@@ -32,11 +20,7 @@ export function PluginsIcon({ id, color }) {
 					<path d="M3 9v14l12 6V15L3 9Z" />
 					<path d="M27 9v14l-12 6V15l12-6Z" />
 				</g>
-				<path
-					d="M11 4h8v2l6 3-10 6L5 9l6-3V4Z"
-					fillOpacity={0.5}
-					className="fill-[var(--icon-background)]"
-				/>
+				<path d="M11 4h8v2l6 3-10 6L5 9l6-3V4Z" fillOpacity={0.5} className="fill-[var(--icon-background)]" />
 				<g
 					className="stroke-[color:var(--icon-foreground)]"
 					strokeWidth={2}
@@ -59,5 +43,5 @@ export function PluginsIcon({ id, color }) {
 				/>
 			</DarkMode>
 		</>
-	)
+	);
 }