+++ packages/sdk/src/index.ts
@@ -1,5 +1,6 @@
 export type { CodeGenerationConfig, ResolvedWunderGraphConfig } from './configure';
 export type {
+	WunderGraphOperationsConfig,
 	BaseOperationConfiguration,
 	QueryConfiguration,
 	MutationConfiguration,
@@ -33,3 +34,4 @@ export {
 export { TypeScriptClient } from './codegen/templates/typescript/client';
 export { PlaceHolder } from './configure/variables';
 export { EnvironmentVariable } from './configure/variables';
+export { configureWunderGraphGeneration } from './configure/codegeneration';