+++ packages/testsuite/apps/http-proxy/.wundergraph/wundergraph.server.ts
@@ -1,0 +1,8 @@
+import { configureWunderGraphServer } from '@wundergraph/sdk/server';
+
+export default configureWunderGraphServer(() => ({
+	hooks: {
+		queries: {},
+		mutations: {},
+	},
+}));